+++ tests/conftest.py
@@ -463,6 +463,57 @@ def sample_move_architecture():
     }
 
 
+@pytest.fixture
+def sample_dynamic_quantum_architecture():
+    return {
+        'calibration_set_id': 'cd4dd889-b88b-4370-ba01-eb8262ad9c53',
+        'qubits': ['QB1', 'QB2', 'QB3'],
+        'computational_resonators': ['COMP_R'],
+        'gates': {
+            'prx': {
+                'implementations': {
+                    'drag_gaussian': {
+                        'loci': [['QB1'], ['QB2'], ['QB3']],
+                    },
+                    'drag_crf': {
+                        'loci': [['QB1'], ['QB2'], ['QB3']],
+                    },
+                },
+                'default_implementation': 'drag_gaussian',
+                'override_default_implementation': {('QB3',): 'drag_crf'},
+            },
+            'cz': {
+                'implementations': {
+                    'tgss': {
+                        'loci': [['QB1', 'COMP_R'], ['QB3', 'COMP_R']],
+                    },
+                    'crf': {
+                        'loci': [['QB1', 'COMP_R']],
+                    },
+                },
+                'default_implementation': 'tgss',
+                'override_default_implementation': {},
+            },
+            'move': {
+                'implementations': {
+                    'tgss_crf': {
+                        'loci': [['QB2', 'COMP_R']],
+                    },
+                },
+                'default_implementation': 'tgss_crf',
+                'override_default_implementation': {},
+            },
+            'measure': {
+                'implementations': {
+                    'constant': {'loci': [['QB1'], ['QB2'], ['QB3']]},
+                },
+                'default_implementation': 'constant',
+                'override_default_implementation': {},
+            },
+        },
+    }
+
+
 class MockTextResponse:
     def __init__(self, status_code: int, text: str, history: Optional[list[Response]] = None):
         self.status_code = status_code
@@ -523,6 +574,11 @@ def quantum_architecture_success(sample_quantum_architecture) -> MockJsonRespons
 @pytest.fixture()
 def move_architecture_success(sample_move_architecture) -> MockJsonResponse:
     return MockJsonResponse(200, sample_move_architecture)
+
+
+@pytest.fixture()
+def dynamic_quantum_architecture_success(sample_dynamic_quantum_architecture) -> MockJsonResponse:
+    return MockJsonResponse(200, sample_dynamic_quantum_architecture)
 
 
 @pytest.fixture()